Test: Service단 테스트 수정
Service단 Mockito 사용하는 것으로 수정 완료

See Also: #19
