+++ src/test/java/com/idolcollector/idolcollector/service/NestedCommentServiceTest.java
@@ -1,7 +1,12 @@
 package com.idolcollector.idolcollector.service;
 
+import com.idolcollector.idolcollector.EntityMaker;
+import com.idolcollector.idolcollector.advice.exception.CAccessDeniedException;
 import com.idolcollector.idolcollector.domain.comment.Comment;
 import com.idolcollector.idolcollector.domain.comment.CommentRepository;
+import com.idolcollector.idolcollector.domain.like.LikeType;
+import com.idolcollector.idolcollector.domain.like.Likes;
+import com.idolcollector.idolcollector.domain.like.LikesRepository;
 import com.idolcollector.idolcollector.domain.member.Member;
 import com.idolcollector.idolcollector.domain.member.MemberRepository;
 import com.idolcollector.idolcollector.domain.member.MemberRole;
@@ -16,6 +21,11 @@ import com.idolcollector.idolcollector.web.dto.nestedcomment.NestedCommentSaveRe
 import com.idolcollector.idolcollector.web.dto.nestedcomment.NestedCommentUpdateRequestDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
@@ -23,120 +33,217 @@ import org.springframework.transaction.annotation.Transactional;
 import javax.servlet.http.HttpSession;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 
-import static org.assertj.core.api.Assertions.*;
+import static com.idolcollector.idolcollector.EntityMaker.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.*;
 
-@SpringBootTest
-@Transactional
+@ExtendWith(MockitoExtension.class)
 class NestedCommentServiceTest {
 
-    @Autowired NestedCommentService nestedCommentService;
+    @InjectMocks
+    NestedCommentService nestedCommentService;
 
-    @Autowired NestedCommentRepository nestedCommentRepository;
-    @Autowired MemberRepository memberRepository;
-    @Autowired CommentRepository commentRepository;
-    @Autowired PostRepository postRepository;
-    @Autowired NoticeRepository noticeRepository;
-    @Autowired HttpSession httpSession;
-
-    @BeforeEach
-    void before() {
-        Member member = new Member(MemberRole.USER, "qqqqeeeererwr#@#wr13", "email", "1111", "steve", "dsfsdfdsfdsf", LocalDateTime.now());
-        memberRepository.save(member);
-        Post post = new Post(member, "title", "conten", "ste", "ori");
-        postRepository.save(post);
-        Comment comment = new Comment(member, post, "content");
-        commentRepository.save(comment);
-    }
-
-    @Test
-    void 저장_조회() {
-
-        // Given
-        Member member = memberRepository.findByNickName("qqqqeeeererwr#@#wr13").get();
-        httpSession.setAttribute("loginMember", member.getId());
-        Comment comment = commentRepository.findAllByMemberId(member.getId()).get(0);
-
-        NestedCommentSaveRequestDto form = new NestedCommentSaveRequestDto(comment.getId(), "nest comment test");
-
-        // When
-        Long save = nestedCommentService.save(form);
-        NestedCommentResponseDto nComment = nestedCommentService.findById(save);
-
-        // Then
-        assertThat(nComment.getContent()).isEqualTo("nest comment test");
-
-        List<Notice> notices = member.getNotices();
-        assertThat(notices.size()).isEqualTo(1);
-    }
+    @Mock NestedCommentRepository nestedCommentRepository;
+    @Mock MemberRepository memberRepository;
+    @Mock CommentRepository commentRepository;
+    @Mock PostRepository postRepository;
+    @Mock NoticeRepository noticeRepository;
+    @Mock HttpSession httpSession;
+    @Mock LikesRepository likesRepository;
 
 
     @Test
-    void 수정() {
-
+    void 저장_성공() {
         // Given
-        Member member = memberRepository.findByNickName("qqqqeeeererwr#@#wr13").get();
-        httpSession.setAttribute("loginMember", member.getId());
-        Comment comment = commentRepository.findAllByMemberId(member.getId()).get(0);
-        NestedComment saved = nestedCommentRepository.save(new NestedComment(member, comment, "nest comment test"));
+        Member sessionMember = generateMember();
+        sessionMember.test(1L);
+        Member author = generateMember();
+        Post post = generatePost(generateMember());
+        Comment comment = post.getComments().get(0);
+        NestedComment nestedComment = generateNComment(comment, author);
 
-        NestedCommentUpdateRequestDto form = new NestedCommentUpdateRequestDto(saved.getId(), "updated nest comment test");
+        Optional<Member> sessionOp = Optional.of(sessionMember);
+        Optional<Comment> commentOp = Optional.of(comment);
+
+        doReturn(commentOp).when(commentRepository).findById(5L);
+        doReturn(sessionOp).when(memberRepository).findById(1L);
+        doReturn(1L).when(httpSession).getAttribute(any());
+        doReturn(new Notice()).when(noticeRepository).save(any(Notice.class));
+        doReturn(nestedComment).when(nestedCommentRepository).save(any());
+
+        NestedCommentSaveRequestDto form = new NestedCommentSaveRequestDto(5L, "nComment content");
 
         // When
-        Long updatedId = nestedCommentService.update(form);
-        NestedComment result = nestedCommentRepository.findById(updatedId).get();
+        nestedCommentService.save(form);
 
-        // Then
-        assertThat(result.getContent()).isEqualTo("updated nest comment test");
-        assertThat(result.getCreateDate()).isNotEqualTo(result.getModifyDate());
+        // Verify
+        verify(nestedCommentRepository, times(1)).save(any());
     }
 
 
     @Test
-    void 삭제() {
-
+    void 수정_성공() {
         // Given
-        Member member = memberRepository.findByNickName("qqqqeeeererwr#@#wr13").get();
-        httpSession.setAttribute("loginMember", member.getId());
-        Comment comment = commentRepository.findAllByMemberId(member.getId()).get(0);
+        Member sessionMember = generateMember();
+        sessionMember.test(1L);
+        Member author = generateMember();
+        author.test(1L);
+        Post post = generatePost(generateMember());
+        Comment comment = post.getComments().get(0);
+        NestedComment nestedComment = generateNComment(comment, author);
 
-        //Long id = nestedCommentService.save(new NestedCommentSaveRequestDto(comment.getId(), "nest comment test"));
+        Optional<Member> sessionOp = Optional.of(sessionMember);
+        Optional<NestedComment> nCommentOp = Optional.of(nestedComment);
 
-        //NestedComment saved = nestedCommentRepository.save(new NestedComment(member, comment, "nest comment test"));
+        doReturn(nCommentOp).when(nestedCommentRepository).findById(4L);
+        doReturn(sessionOp).when(memberRepository).findById(1L);
+        doReturn(1L).when(httpSession).getAttribute(any());
+
+        NestedCommentUpdateRequestDto form = new NestedCommentUpdateRequestDto(4L, "updated n comment");
 
         // When
-        List<NestedComment> result = comment.getNComment();
-
-        for (NestedComment nestedComment : result) {
-            System.out.println("22222222 = " + nestedComment.getContent());
-        }
-
-        //nestedCommentService.delete(id);
-
-        for (NestedComment nestedComment : result) {
-            System.out.println("3333333 " + nestedComment.getContent());
-        }
+        nestedCommentService.update(form);
 
         // Then
-        assertThat(result.size()).isEqualTo(0);
+        assertThat(nestedComment.getContent()).isEqualTo("updated n comment");
     }
 
 
     @Test
-    void 좋아요() {
-
+    void 수정_실패() {
         // Given
-        Member member = memberRepository.findByNickName("qqqqeeeererwr#@#wr13").get();
-        httpSession.setAttribute("loginMember", member);
-        Comment comment = commentRepository.findAllByMemberId(member.getId()).get(0);
-        NestedComment nComment = nestedCommentRepository.save(new NestedComment(member, comment, "nest comment test"));
+        Member sessionMember = generateMember();
+        sessionMember.test(1L);
+        Member author = generateMember();
+        author.test(2L);
+        Post post = generatePost(generateMember());
+        Comment comment = post.getComments().get(0);
+        NestedComment nestedComment = generateNComment(comment, author);
+
+        Optional<Member> sessionOp = Optional.of(sessionMember);
+        Optional<NestedComment> nCommentOp = Optional.of(nestedComment);
+
+        doReturn(nCommentOp).when(nestedCommentRepository).findById(4L);
+        doReturn(sessionOp).when(memberRepository).findById(1L);
+        doReturn(1L).when(httpSession).getAttribute(any());
+
+        NestedCommentUpdateRequestDto form = new NestedCommentUpdateRequestDto(4L, "updated n comment");
 
         // When
-        nComment.addLike();
 
-        NestedComment find = nestedCommentRepository.findById(nComment.getId()).get();
+        assertThatThrownBy(() -> {
+            nestedCommentService.update(form);
+        }).isInstanceOf(CAccessDeniedException.class);
 
-        // Then
-        assertThat(find.getLikes()).isEqualTo(1);
+    }
+
+
+    @Test
+    void 삭제_성공() {
+        // Given
+        Member sessionMember = generateMember();
+        sessionMember.test(1L);
+        Member author = generateMember();
+        author.test(1L);
+        Post post = generatePost(generateMember());
+        Comment comment = post.getComments().get(0);
+        NestedComment nestedComment = generateNComment(comment, author);
+
+        Optional<Member> sessionOp = Optional.of(sessionMember);
+        Optional<NestedComment> nCommentOp = Optional.of(nestedComment);
+
+        doReturn(nCommentOp).when(nestedCommentRepository).findById(4L);
+        doReturn(sessionOp).when(memberRepository).findById(1L);
+        doReturn(1L).when(httpSession).getAttribute(any());
+        doNothing().when(nestedCommentRepository).delete(nestedComment);
+
+        // When
+        nestedCommentService.delete(4L);
+
+        // Verify
+        verify(nestedCommentRepository, times(1)).delete(any());
+    }
+
+
+    @Test
+    void 삭제_실패() {
+        // Given
+        Member sessionMember = generateMember();
+        sessionMember.test(1L);
+        Member author = generateMember();
+        author.test(2L);
+        Post post = generatePost(generateMember());
+        Comment comment = post.getComments().get(0);
+        NestedComment nestedComment = generateNComment(comment, author);
+
+        Optional<Member> sessionOp = Optional.of(sessionMember);
+        Optional<NestedComment> nCommentOp = Optional.of(nestedComment);
+
+        doReturn(nCommentOp).when(nestedCommentRepository).findById(4L);
+        doReturn(sessionOp).when(memberRepository).findById(1L);
+        doReturn(1L).when(httpSession).getAttribute(any());
+
+        // When
+        assertThatThrownBy(() -> {
+            nestedCommentService.delete(4L);
+        }).isInstanceOf(CAccessDeniedException.class);
+    }
+
+
+    @Test
+    void 좋아요_성공() {
+        // Given
+        Member sessionMember = generateMember();
+        sessionMember.test(1L);
+        Member author = generateMember();
+        author.test(1L);
+        Post post = generatePost(generateMember());
+        Comment comment = post.getComments().get(0);
+        NestedComment nestedComment = generateNComment(comment, author);
+
+        Optional<Member> sessionOp = Optional.of(sessionMember);
+        Optional<NestedComment> nCommentOp = Optional.of(nestedComment);
+
+        doReturn(nCommentOp).when(nestedCommentRepository).findById(4L);
+        doReturn(sessionOp).when(memberRepository).findById(1L);
+        doReturn(1L).when(httpSession).getAttribute(any());
+        doReturn(new Likes()).when(likesRepository).save(any());
+
+        // When
+        nestedCommentService.like(4L);
+
+        // Verify
+        verify(likesRepository, times(1)).save(any());
+    }
+
+
+    @Test
+    void 좋아요_실패() {
+        // Given
+        Member sessionMember = generateMember();
+        sessionMember.test(1L);
+        Member author = generateMember();
+        author.test(1L);
+        Post post = generatePost(generateMember());
+        Comment comment = post.getComments().get(0);
+        NestedComment nestedComment = generateNComment(comment, author);
+
+        Optional<Member> sessionOp = Optional.of(sessionMember);
+        Optional<NestedComment> nCommentOp = Optional.of(nestedComment);
+        Optional<Likes> likesOp = Optional.of(new Likes());
+
+        doReturn(nCommentOp).when(nestedCommentRepository).findById(4L);
+        doReturn(sessionOp).when(memberRepository).findById(1L);
+        doReturn(1L).when(httpSession).getAttribute(any());
+        doReturn(likesOp).when(likesRepository).findLikeByMemberIdPostId(any(), any(), any());
+
+        // When
+        assertThatThrownBy(() -> {
+            nestedCommentService.like(4L);
+        }).isInstanceOf(IllegalArgumentException.class);
     }
 }
