+++ src/test/java/com/idolcollector/idolcollector/service/PostServiceTest.java
@@ -1,7 +1,12 @@
 package com.idolcollector.idolcollector.service;
 
+import com.idolcollector.idolcollector.EntityMaker;
+import com.idolcollector.idolcollector.advice.exception.CAccessDeniedException;
+import com.idolcollector.idolcollector.advice.exception.CPostNotFoundException;
 import com.idolcollector.idolcollector.domain.comment.Comment;
 import com.idolcollector.idolcollector.domain.comment.CommentRepository;
+import com.idolcollector.idolcollector.domain.like.LikeType;
+import com.idolcollector.idolcollector.domain.like.Likes;
 import com.idolcollector.idolcollector.domain.like.LikesRepository;
 import com.idolcollector.idolcollector.domain.member.Member;
 import com.idolcollector.idolcollector.domain.member.MemberRepository;
@@ -10,6 +15,7 @@ import com.idolcollector.idolcollector.domain.nestedcomment.NestedComment;
 import com.idolcollector.idolcollector.domain.nestedcomment.NestedCommentRepository;
 import com.idolcollector.idolcollector.domain.notice.Notice;
 import com.idolcollector.idolcollector.domain.notice.NoticeRepository;
+import com.idolcollector.idolcollector.domain.notice.NoticeType;
 import com.idolcollector.idolcollector.domain.post.Post;
 import com.idolcollector.idolcollector.domain.post.PostRepository;
 import com.idolcollector.idolcollector.domain.posttag.PostTag;
@@ -21,15 +27,25 @@ import com.idolcollector.idolcollector.domain.tag.TagRepository;
 import com.idolcollector.idolcollector.domain.trending.Trending;
 import com.idolcollector.idolcollector.domain.trending.TrendingRepository;
 import com.idolcollector.idolcollector.domain.trending.TrendingType;
+import com.idolcollector.idolcollector.file.FileStore;
+import com.idolcollector.idolcollector.web.dto.comment.CommentResponseDto;
 import com.idolcollector.idolcollector.web.dto.post.HomePostListResponseDto;
 import com.idolcollector.idolcollector.web.dto.post.PostResponseDto;
 import com.idolcollector.idolcollector.web.dto.post.PostSaveRequestDto;
 import com.idolcollector.idolcollector.web.dto.post.PostUpdateRequestDto;
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.mock.web.MockMultipartFile;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -40,297 +56,253 @@ import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
+import static com.idolcollector.idolcollector.EntityMaker.*;
 import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.InstanceOfAssertFactories.OPTIONAL;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.*;
 
-@SpringBootTest
-@Transactional
+@ExtendWith(MockitoExtension.class)
 class PostServiceTest {
 
-    @Autowired
+    @InjectMocks
     PostService postService;
 
-    @Autowired PostRepository postRepository;
-    @Autowired MemberRepository memberRepository;
-    @Autowired TagRepository tagRepository;
-    @Autowired PostTagRepository postTagRepository;
-    @Autowired CommentRepository commentRepository;
-    @Autowired NestedCommentRepository nestedCommentRepository;
-    @Autowired ScrapRepository scrapRepository;
-    @Autowired LikesRepository likesRepository;
-    @Autowired TrendingRepository trendingRepository;
-    @Autowired NoticeRepository noticeRepository;
-    @Autowired HttpSession httpSession;
+    @Mock TagService tagService;
 
-    @BeforeEach
-    void before() {
-        Member member = new Member(MemberRole.USER, "qqqqeeeererwr#@#wr13", "email", "1111", "steve", "dsfsdfdsfdsf", LocalDateTime.now());
-        memberRepository.save(member);
-    }
+    @Mock PostRepository postRepository;
+    @Mock MemberRepository memberRepository;
+    @Mock TagRepository tagRepository;
+    @Mock ScrapRepository scrapRepository;
+    @Mock LikesRepository likesRepository;
+    @Mock TrendingRepository trendingRepository;
+    @Mock NoticeRepository noticeRepository;
 
-    @Test
-    void 페이지_리스트받아오기() {
-
-        // Given
-        Member member = memberRepository.findByNickName("qqqqeeeererwr#@#wr13").get();
-
-        Post post = new Post(member, "title", "conten", "ste", "ori");
-        Post p1 = postRepository.save(post);
-
-        for (int i = 0; i < 250; i++) {
-            postRepository.save(new Post(member, "test" + i, "content", "ste", "ori"));
-        }
-
-        for (int i = 0; i < 30; i++) {
-            Post save = postRepository.save(new Post(member, "viewd" + i, "content", "ste", "ori"));
-            trendingRepository.save(new Trending(save, TrendingType.LIKE));
-        }
-
-        Post post2 = new Post(member, "title2", "conten2", "ste", "ori");
-        Post p2 = postRepository.save(post2);
-
-
-        Trending p1T1 = new Trending(p1, TrendingType.VIEW);
-        Trending p1T2 = new Trending(p1, TrendingType.VIEW);
-        Trending p1T3 = new Trending(p1, TrendingType.LIKE);
-        Trending p1T4 = new Trending(p1, TrendingType.COMMENT);
-        Trending p1T5 = new Trending(p1, TrendingType.COMMENT);
-        Trending p1T6 = new Trending(p1, TrendingType.COMMENT);
-        Trending p1T7 = new Trending(p1, TrendingType.VIEW);
-
-        trendingRepository.save(p1T1);
-        trendingRepository.save(p1T2);
-        trendingRepository.save(p1T3);
-        trendingRepository.save(p1T4);
-        trendingRepository.save(p1T5);
-        trendingRepository.save(p1T6);
-        trendingRepository.save(p1T7);
-
-        Trending p2T1 = new Trending(p2, TrendingType.VIEW);
-        Trending p2T2 = new Trending(p2, TrendingType.VIEW);
-        Trending p2T3 = new Trending(p2, TrendingType.VIEW);
-
-        trendingRepository.save(p2T1);
-        trendingRepository.save(p2T2);
-        trendingRepository.save(p2T3);
-
-        // When
-        List<HomePostListResponseDto> list = postService.scorePostList(5);
-
-        for (HomePostListResponseDto homePostListResponseDto : list) {
-            System.out.println("homePostListResponseDto.getTitle() = " + homePostListResponseDto.getTitle());
-        }
-    }
-
+    @Mock FileStore fileStore;
+    @Mock HttpSession httpSession;
 
     @Test
     void 생성() throws IOException {
-
         // Given
-        Member member = memberRepository.findByNickName("qqqqeeeererwr#@#wr13").get();
-
-        httpSession.setAttribute("loginMember", member.getId());
-
+        Post post = generatePost(generateMember());
         List<String> tags = new ArrayList<>();
-        tags.add("aSDasdawDW");
-        tags.add("DFeafsfefewff");
-        tags.add("WEFSADVsdfafaWEFASF");
+        tags.add("tag1");
+        tags.add("tag2");
 
-        // 사진 파일
-        File img = new File("./src/test/resources/test.png");
-        MultipartFile mf = new MockMultipartFile("image","test.png", "img",new FileInputStream(img));
+        Optional<Member> op = Optional.of(generateMember());
 
-        PostSaveRequestDto form = new PostSaveRequestDto(
-                "예쁜이하늬",
-                "이하늬 예쁘다 헤헤",
-                mf,
-                tags);
+        doReturn(1L).when(httpSession).getAttribute(any());
+        doReturn(op).when(memberRepository).findById(1L);
+        doReturn(generateUploadFile()).when(fileStore).storeFile(any(MultipartFile.class));
+        doReturn(post).when(postRepository).save(any(Post.class));
+        doNothing().when(tagService).createPostTag(tags ,post);
 
         // When
-        Long postId = postService.create(form);
+        File img = new File("./src/test/resources/test.png");
+        MockMultipartFile mf = new MockMultipartFile("image","test.png", "img", new FileInputStream(img));
+        PostSaveRequestDto form = new PostSaveRequestDto("제목", "카드 내용", mf, tags);
 
-        // Then
-        Post findPost = postRepository.findById(postId).get();
+        postService.create(form);
 
-        List<PostTag> findPostTags = postTagRepository.findAllByPostId(findPost.getId());
-
-        assertThat(findPost.getId()).isEqualTo(postId);
-        assertThat(findPostTags.size()).isEqualTo(3);
+        //verify
+        verify(memberRepository, times(1)).findById(any());
+        verify(fileStore, times(1)).storeFile(any());
+        verify(postRepository, times(1)).save(any());
+        verify(tagService, times(1)).createPostTag(any(), any());
     }
 
     @Test
     void 상세보기() {
-
         // Given
-        Member member = memberRepository.findByNickName("qqqqeeeererwr#@#wr13").get();
-        httpSession.setAttribute("loginMember", member.getId());
-
-        Post post = postRepository.save(new Post(member, "title", "content", "storeFilename", "oriFileName"));
-        Comment comment = commentRepository.save(new Comment(member, post, "content"));
-        Comment comment2 = commentRepository.save(new Comment(member, post, "content2"));
-
-        NestedComment nComment = nestedCommentRepository.save(new NestedComment(member, comment, "test"));
-        NestedComment nComment2 = nestedCommentRepository.save(new NestedComment(member, comment, "test"));
+        Post post = generatePost(generateMember());
+        Member sessionMember = generateMember();
 
         // When
+        doReturn(1L).when(httpSession).getAttribute(any());
+        doReturn(Optional.of(sessionMember)).when(memberRepository).findById(1L);
+        doReturn(Optional.of(post)).when(postRepository).findById(any());
+        doReturn(Optional.empty()).when(likesRepository).findLikeByMemberIdPostId(any(), any(), any());
+        doReturn(Optional.of(new Scrap(sessionMember, post))).when(scrapRepository).findScrapByMemberIdPostId(any(), any());
+
         PostResponseDto detail = postService.detail(post.getId());
 
         // Then
-        assertThat(detail.getAuthorId()).isEqualTo(member.getId());
-        assertThat(detail.getContent()).isEqualTo(post.getContent());
-        assertThat(detail.getComments().get(0).getContent()).isEqualTo(comment.getContent());
-        assertThat(detail.getComments().get(0).getNestedComments().get(0).getContent()).isEqualTo(nComment.getContent());
-        assertThat(post.getViews()).isEqualTo(1);
+        assertThat(detail.getAuthorNickName()).isEqualTo("testMember");
+        assertThat(detail.getTitle()).isEqualTo("test card");
+        assertThat(detail.getStoreFileName()).isEqualTo("ste");
+        assertThat(detail.getComments().size()).isEqualTo(4);
+        assertThat(detail.getComments().get(0).getContent()).contains("comment");
+        assertThat(detail.getTags().size()).isEqualTo(2);
+        assertThat(detail.getTags().get(0).getName()).contains("tag");
+        assertThat(detail.isDidScrap()).isTrue();
     }
 
     @Test
     void 수정() throws IOException {
-
         // Given
-        Member member = memberRepository.findByNickName("qqqqeeeererwr#@#wr13").get();
+        Post post = generatePost(generateMember());
+        Member sessionMember = generateMember();
 
-        httpSession.setAttribute("loginMember", member.getId());
+        doReturn(1L).when(httpSession).getAttribute(any());
+        doReturn(Optional.of(post)).when(postRepository).findById(any());
+        doReturn(Optional.of(post.getMember())).when(memberRepository).findById(1L);
 
+        doNothing().when(tagRepository).deleteAllByPostId(any());
+        List<PostTag> postTags = post.getPostTags();
+        while (postTags.size() > 0) postTags.remove(0);
 
-        List<String> tags = new ArrayList<>();
-        tags.add("이하늬");
-        tags.add("예쁨");
-        tags.add("BEautiful");
-
-        // 사진 파일
-        File img = new File("./src/test/resources/test.png");
-        MultipartFile mf = new MockMultipartFile("image","test.png", "img",new FileInputStream(img));
-
-        PostSaveRequestDto saveForm = new PostSaveRequestDto(
-                "예쁜이하늬",
-                "이하늬 예쁘다 lol 헤헤",
-                mf,
-                tags);
-
-        Long postId = postService.create(saveForm);
+        doNothing().when(tagService).createPostTag(any(), any());
+        for (int i = 0; i < 3; i++) postTags.add(new PostTag(post, new Tag("updated tag"+i)));
 
         // When
-        List<String> updateTags = new ArrayList<>();
-        updateTags.add("이하늬");
-        updateTags.add("타짜2");
-
-        PostUpdateRequestDto form = new PostUpdateRequestDto(postId,
-                "2하2",
-                "타짜에서 나왔구나.",
-                updateTags);
-
-        Long updateId = postService.update(form);
-
+        PostUpdateRequestDto form = new PostUpdateRequestDto(post.getId(), "updated title", "updated description", new ArrayList<String>());
+        postService.update(form);
 
         // Then
-        Post updated = postRepository.findById(updateId).get();
+        assertThat(post.getTitle()).isEqualTo("updated title");
+        assertThat(post.getContent()).isEqualTo("updated description");
+        assertThat(post.getPostTags().size()).isEqualTo(3);
+    }
 
-        assertThat(updated.getTitle()).isEqualTo("2하2");
-        assertThat(updated.getContent()).isEqualTo("타짜에서 나왔구나.");
-        assertThat(updated.getModifyDate()).isNotEqualTo(updated.getCreateDate());
-        assertThat(updated.getOriFileName()).isEqualTo("test.png");
-        assertThat(updated.getStoreFileName()).isNotNull();
+    @Test
+    void 수정_실패_작성자_아님() throws IOException {
+        // Given
+        Post post = generatePost(generateMember());
+        Member sessionMember = generateMember();
+        post.getMember().test(1L); sessionMember.test(2L);
 
-        List<Tag> findTags = tagRepository.findAllByPostId(updated.getId());
+        doReturn(1L).when(httpSession).getAttribute(any());
+        doReturn(Optional.of(post)).when(postRepository).findById(any());
+        doReturn(Optional.of(sessionMember)).when(memberRepository).findById(1L);
 
-        assertThat(findTags.size()).isEqualTo(2);
-        assertThat(findTags.get(0).getName()).isEqualTo("이하늬");
+        // Then
+        PostUpdateRequestDto form = new PostUpdateRequestDto(post.getId(), "updated title", "updated description", new ArrayList<String>());
+        assertThatThrownBy(() -> {
+            postService.update(form);
+        }).isInstanceOf(CAccessDeniedException.class);
+
+        // Verify
+        verify(tagRepository, times(0)).deleteAllByPostId(any());
+        verify(tagService, times(0)).createPostTag(any(), any());
     }
 
     @Test
     void 삭제() {
-
         // Given
-        Member member = memberRepository.findByNickName("qqqqeeeererwr#@#wr13").get();
-        Post post = postRepository.save(new Post(member, "title", "content", "storeFilename", "oriFileName"));
-        Comment comment = commentRepository.save(new Comment(member, post, "content"));
-        Comment comment2 = commentRepository.save(new Comment(member, post, "content2"));
+        Post post = generatePost(generateMember());
+        Member sessionMember = generateMember();
 
-        NestedComment nComment = nestedCommentRepository.save(new NestedComment(member, comment, "test"));
-        NestedComment nComment2 = nestedCommentRepository.save(new NestedComment(member, comment, "test"));
+        doReturn(1L).when(httpSession).getAttribute(any());
+        doReturn(Optional.of(post)).when(postRepository).findById(any());
+        doReturn(Optional.of(post.getMember())).when(memberRepository).findById(1L);
+        doNothing().when(postRepository).delete(any(Post.class));
 
         // When
-        postRepository.delete(post);
+        postService.delete(1L);
 
-        // Then
-        assertThatThrownBy(() -> {
-            postRepository.findById(post.getId())
-                    .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 게시글 입니다."));
-        }).isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("존재하지 않는 게시글 입니다.");
-
-        assertThatThrownBy(() -> {
-            commentRepository.findById(comment.getId())
-                    .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 댓글 입니다."));
-        }).isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("존재하지 않는 댓글 입니다.");
-
-        assertThatThrownBy(() -> {
-            commentRepository.findById(comment2.getId())
-                    .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 댓글 입니다."));
-        }).isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("존재하지 않는 댓글 입니다.");
-
-        assertThatThrownBy(() -> {
-            commentRepository.findById(nComment.getId())
-                    .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 대댓글 입니다."));
-        }).isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("존재하지 않는 대댓글 입니다.");
-
-        assertThatThrownBy(() -> {
-            commentRepository.findById(nComment2.getId())
-                    .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 대댓글 입니다."));
-        }).isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("존재하지 않는 대댓글 입니다.");
-
+        // Verify
+        verify(memberRepository, times(1)).findById(any());
+        verify(postRepository, times(1)).findById(any());
+        verify(postRepository, times(1)).delete(any());
     }
 
     @Test
-    void 좋아요() {
-
+    void 삭제_실패_작성자_아님() {
         // Given
-        Member member = memberRepository.findByNickName("qqqqeeeererwr#@#wr13").get();
-        httpSession.setAttribute("loginMember", member.getId());
+        Post post = generatePost(generateMember());
+        Member sessionMember = generateMember();
+        post.getMember().test(1L); sessionMember.test(2L);
 
-        Post post = postRepository.save(new Post(member, "title", "content", "storeFilename", "oriFileName"));
-
-        // When
-        postService.like(post.getId());
+        doReturn(1L).when(httpSession).getAttribute(any());
+        doReturn(Optional.of(post)).when(postRepository).findById(1L);
+        doReturn(Optional.of(sessionMember)).when(memberRepository).findById(1L);
 
         // Then
-        assertThat(post.getLikes()).isEqualTo(1);
+        assertThatThrownBy(() -> {
+            postService.delete(1L);
+        }).isInstanceOf(CAccessDeniedException.class);
 
-        List<Notice> notices = member.getNotices();
-        assertThat(notices.size()).isEqualTo(1);
+        // Verify
+        verify(memberRepository, times(1)).findById(any());
+        verify(postRepository, times(1)).findById(any());
+        verify(postRepository, times(0)).delete(any(Post.class));
+    }
 
 
+    @Test
+    void 좋아요_성공() {
+        // Given
+        Post post = generatePost(generateMember());
+        Member sessionMember = generateMember();
+
+        doReturn(1L).when(httpSession).getAttribute(any());
+        doReturn(Optional.of(post)).when(postRepository).findById(any());
+        doReturn(Optional.of(post.getMember())).when(memberRepository).findById(1L);
+        doReturn(Optional.empty()).when(likesRepository).findLikeByMemberIdPostId(post.getId(), sessionMember.getId(), LikeType.POST);
+        doReturn(new Likes()).when(likesRepository).save(any(Likes.class));
+        doReturn(new Notice()).when(noticeRepository).save(any(Notice.class));
+        doReturn(new Trending()).when(trendingRepository).save(any(Trending.class));
+
+        // When
+        int likes = postService.like(post.getId());
+
+        // Then
+        assertThat(likes).isEqualTo(1);
+
+        // Verify
+        verify(likesRepository, times(1)).save(any());
+        verify(noticeRepository, times(1)).save(any());
+        verify(trendingRepository, times(1)).save(any());
+    }
+
+    @Test
+    void 좋아요_실패_중복() {
+        // Given
+        Post post = generatePost(generateMember());
+        Member sessionMember = generateMember();
+
+        doReturn(1L).when(httpSession).getAttribute(any());
+        doReturn(Optional.of(post)).when(postRepository).findById(any());
+        doReturn(Optional.of(post.getMember())).when(memberRepository).findById(1L);
+        doReturn(Optional.of(new Likes())).when(likesRepository).findLikeByMemberIdPostId(post.getId(), sessionMember.getId(), LikeType.POST);
+
+        // Then
+        assertThatThrownBy(() -> {
+            postService.like(post.getId());
+        }).isInstanceOf(IllegalArgumentException.class);
+
+        // Verify
+        verify(likesRepository, times(0)).save(any());
+        verify(noticeRepository, times(0)).save(any());
+        verify(trendingRepository, times(0)).save(any());
     }
 
     @Test
     void 스크랩() {
-
         // Given
-        Member member = new Member(MemberRole.USER, "nick", "email", "1111", "david", "dsfsdfdsfdsf", LocalDateTime.now());
-        memberRepository.save(member);
+        Post post = generatePost(generateMember());
+        Member sessionMember = generateMember();
 
-        httpSession.setAttribute("loginMember", member.getId());
+        doReturn(1L).when(httpSession).getAttribute(any());
+        doReturn(Optional.of(post)).when(postRepository).findById(any());
+        doReturn(Optional.of(sessionMember)).when(memberRepository).findById(1L);
+        doReturn(new Scrap()).when(scrapRepository).save(any());
 
-        Post post = postRepository.save(new Post(member, "title", "content", "storeFilename", "oriFileName"));
+        doReturn(new Notice()).when(noticeRepository).save(any());
+        doReturn(new Trending()).when(trendingRepository).save(any());
 
         // When
-        Long id = postService.scrap(post.getId());
+        Long scrap = postService.scrap(post.getId());
 
-        // Then
-        Scrap scrap = scrapRepository.findScrapByMemberIdPostId(post.getId(), member.getId()).get();
-
-        assertThat(scrap.getPost().getTitle()).isEqualTo("title");
-        assertThat(scrap.getMember().getNickName()).isEqualTo("nick");
-
-        List<Notice> notices = member.getNotices();
-        assertThat(notices.size()).isEqualTo(1);
-
+        // Verify
+        verify(postRepository, times(1)).findById(any());
+        verify(memberRepository, times(1)).findById(any());
+        verify(scrapRepository, times(1)).save(any());
+        verify(noticeRepository, times(1)).save(any());
+        verify(trendingRepository, times(1)).save(any());
     }
-
-
-
 }
